software_side/walkbuddy_reactNative/backend/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
pydantic==2.9.2
opencv-python==4.10.0.84
easyocr==1.7.1
pyttsx3
jiwer==3.0.4
nltk==3.9.1
ultralytics>=8.0
numpy==1.26.4
scikit-image<0.22
gTTs
gradio==3.50.2
scipy==1.13.1
pandas==2.2.2
python-dotenv
